facebook/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
distlib==0.3.8
Django==5.0.1
django-multiupload==0.6.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
oauthlib==3.2.2
mysqlclient==2.2.4
packaging==23.2
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==1.7.1
pyOpenSSL==24.1.0
pyotp==2.9.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.1
zope.interface==6.4